rename: PlaybookPage 이름에 대한 에러 발견
