+++ src/pages/teamroom/Play.tsx
@@ -1,4 +1,4 @@
-import { PlaybookPageLayout } from "./PlaybookPageStyles";
+import { PlaybookPageLayout } from "./PlayStyles";
 
 const PlaybookPage = () => {
   return <PlaybookPageLayout>PlayBook 페이지입니다.</PlaybookPageLayout>;
